pyproject.toml
[tool.poetry]
name = "djapy"
version = "0.1.0"
description = ""
authors = ["wilfred11 <wilfredvancasteren@hotmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0.4"
colorama = "^0.4.6"
requests = "^2.31.0"
django-environ = "^0.11.2"
aiohttp = "^3.9.5"
uvicorn = "^0.29.0"
httpx = "^0.27.0"
psycopg2 = "^2.9.9"
django-rest-framework = "^0.1.0"
django-extensions = "^3.2.3"
django-datatables-view = "^1.20.0"
black = "^24.4.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
